refactor: Menü-Zeilen-Composables nach SettingsMenuItems.kt extrahiert
Reiner Move ohne Verhaltensänderung: EditMenuItem, EditToggleItem,
EditSectionHeader, EditAppAccessSelectorItem, EditIslandAnimationSelectorItem,
EditActionSelectorItem, gestureActionIcon/Label und GestureAppPickerDialog
wandern aus EditConfigMenu.kt in die neue Datei SettingsMenuItems.kt
(private → internal), damit die kommenden Einstellungs-Kategorie-Seiten
dieselben Bausteine nutzen können.
